[EditorWork] move rendering logic to tab
